refactor: rename workspace command to agent
- Remove old half-implemented agent command
- Rename workspace → agent to better reflect purpose
- Update all user-facing text to "agent environment"
- Fix help text to show correct dev options (docker, dagger, native)
- Document refactoring decisions in layer

Following container-use philosophy where agents work in isolated
environments. Internal workspace terminology preserved where accurate.
