+++ src/commands/mod.rs
@@ -8,4 +8,3 @@ pub mod push;
 pub mod test;
 pub mod update;
 pub mod version;
-pub mod workspace;
